requirements.txt
flask==3.0.0
transformers==4.36.0
waitress==2.1.2
numpy==1.24.3
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
python-dotenv==1.0.0